Export type definitions
Hand-written definitions for Mori types are in mori.js.flow. In projects that
use Flow for type-checking and that import Mori, Flow will load those
definitions automatically.

mori-spec.js is modified slightly to signal to Flow that it should type-check
that file. This is to make sure that the type-checker does not infer errors in
any code in the unit tests - that would indicate that there is something wrong
in the type definitions.

types-spec.js has been added to provide certain test cases that should pass
type-checking, to make sure that those cases do pass. types-spec.js includes
type annotations as specially-formatted comments.

flow-bin has been added as a development dependency, and .flowconfig has been
added to configure Flow type checking. Type checking can be run with:

    $ flow check

This will check code in mori-spec.js and in types-spec.js against the
definitions in mori.js.flow.